\name{activePaths}
\alias{activePaths}
\title{Find all active paths in a (partially) directed graph...}
\usage{activePaths(graph, node, nodeNames)
}
\description{Find all active paths in a (partially) directed graph}
\value{a list of active paths with node as its source}
\arguments{\item{graph}{the graph either in one of the package \code{graph} classes, or of class \code{bn} or \code{pcAlgo}}
\item{node}{the source node of the path (index not name)}
\item{nodeNames}{optionally specify node names which can be used to return those instead of indicies}
}

