\name{adjC.allVarInx}
\alias{adjC.allVarInx}
\title{Get all the variable indicies in adjC, both target and condSet...}
\usage{adjC.allVarInx(adjC)
}
\description{Get all the variable indicies in adjC, both target and condSet}
\value{numeric vector (unique values)}
\arguments{\item{adjC}{the adjC list of conditional independence tests for variables "adjacent" to target variable C}
}

