\name{adjC.allVarNames}
\alias{adjC.allVarNames}
\title{Get all the variable names in adjC, both target and condSet...}
\usage{adjC.allVarNames(adjC)
}
\description{Get all the variable names in adjC, both target and condSet}
\value{character vector (unique names)}
\arguments{\item{adjC}{the adjC list of conditional independence tests for variables "adjacent" to target variable C}
}

