\name{adjC.condSetSize}
\alias{adjC.condSetSize}
\title{Returns the total size of conditioning set for adjC (i...}
\usage{adjC.condSetSize(adjC)
}
\description{Returns the total size of conditioning set for adjC (i.e. all variables present in adjC)}
\value{sum of all conditioning set sizes plus size of adjC, i.e. all variables present in adjC}
\arguments{\item{adjC}{the adjC list of conditional independence tests for variables "adjacent" to target variable C}
}

