\name{adjC.targetInx}
\alias{adjC.targetInx}
\title{Get all the targetInx values in adjC...}
\usage{adjC.targetInx(adjC)
}
\description{Get all the targetInx values in adjC}
\value{numeric vector (unique values)}
\arguments{\item{adjC}{the adjC list of conditional independence tests for variables "adjacent" to target variable C}
}

