\name{blockingVariables}
\alias{blockingVariables}
\title{Version of blockingNodes() for DDGraphs...}
\usage{blockingVariables(obj, nodes)
}
\description{Version of blockingNodes() for DDGraphs}
\value{same as blockingNodes(): a list with blocking nodes and minimal length to the target node: target node => blocked by => number of steps}
\arguments{\item{obj}{DDGraph object}
\item{nodes}{the selected nodes}
}

