\name{CITestResultID}
\alias{CITestResultID}
\title{Provide a unique ID composing of target, source and conditioning set (all names)...}
\usage{CITestResultID(citest)
}
\description{Provide a unique ID composing of target, source and conditioning set (all names)}
\value{a character ID}
\arguments{\item{citest}{a CITestResult object}
}

