\name{CITestResultVar}
\alias{CITestResultVar}
\title{Return a string representation of a variable represented with this CITest...}
\usage{CITestResultVar(citest)
}
\description{Return a string representation of a variable represented with this CITest}
\arguments{\item{citest}{an object of class CITestResult}
}

