\name{customPlotPCAlgo}
\alias{customPlotPCAlgo}
\title{Custom plotting for pcalgo}
\usage{customPlotPCAlgo(x, main, labels, colors, ...)
}
\description{Custom plotting function of PC algorithm to have nice highlighting}
\arguments{\item{x}{an object of one of the pcalg classes}
\item{main}{the main title}
\item{labels}{the labels of each of the nodes (doesn't need to be a named vector)}
\item{colors}{the colors we want to assign to each node (doesn't need to be a named vector)}
\item{...}{additional parameters to pass to \code{layoutGraph()}}
}

