\name{datasetName,DDDataSet-method}
\alias{datasetName,DDDataSet-method}
\alias{datasetName}
\title{Dataset name...}
\usage{\S4method{datasetName}{DDDataSet}(obj, ...)
}
\description{Dataset name}
\value{the name of the dataset used in plotting etc}
\arguments{\item{obj}{the DDDataSet object}
\item{...}{unused}
}

