\name{dataType,DDDataSet-method}
\alias{dataType,DDDataSet-method}
\alias{dataType}
\title{Return data type}
\usage{\S4method{dataType}{DDDataSet}(obj, ...)
}
\description{Return the data type ("binary" or "continuous")}
\value{the data type}
\arguments{\item{obj}{the DDDataSet object}
\item{...}{unused}
}

