\name{DDGraph-class}
\alias{DDGraph-class}
\alias{DDGraph}
\title{Direct Dependence Graph class...}
\description{Direct Dependence Graph class}
\details{This class represents one Direct Dependence Graphs (generated by a certain conditional independence test, alpha value, etc).
It contains the original \code{DDDataSet} object from which it stems, the set of parameters, the set of informative 
statistics as well as lists of direct, joint and indirect variables. Finally, it contains the edges needed to 
draw the graph.}
\arguments{\item{dataset}{the DDDataSet object used to make the DDGraph}
\item{params}{the parameters used in making the DDGraph}
\item{stats}{the values of statistics used to make the DDGraph}
\item{direct}{the list of indicies of direct variables}
\item{indirect}{the list of indicies of indirect variables}
\item{joint}{the list of indicies of joint variables}
\item{conditional}{the list of indicies of conditional variables}
\item{conditionalJoint}{the list of indicies of conditionally joint variables}
\item{edges}{the list of edges (type DDGraphEdge) the describe the graph}
}
\section{Methods}{\describe{\item{\code{\link[=initialize,DDGraph-method]{initialize}}}{\code{signature(.Object = "DDGraph")}: Construct new DDGraph object}\item{\code{\link[=names,DDGraph-method]{names}}}{\code{signature(x = "DDGraph")}: Names of different properties that can be accessed with $ operator}\item{\code{\link[=$,DDGraph-method]{$}}}{\code{signature(x = "DDGraph")}: access a property by name}\item{\code{\link[=show,DDGraph-method]{show}}}{\code{signature(object = "DDGraph")}: show method for DDGraph}\item{\code{\link[=plot,DDGraph,missing-method]{plot}}}{\code{signature(x = "DDGraph", y = "missing")}: Plot DDGraphs using RGraphviz}}}

