\name{entropyFromFreq}
\alias{entropyFromFreq}
\title{Calculate entropy from frequencies of observations for discrete data...}
\usage{entropyFromFreq(x)
}
\description{Calculate entropy from frequencies of observations for discrete data}
\value{the entropy in bits}
\arguments{\item{x}{the vector of frequencies, or a pdf of distribution}
}

