\name{foldChangeFromFreq}
\alias{foldChangeFromFreq}
\title{Calculate the fold change when x is of size two (always show it >1)...}
\usage{foldChangeFromFreq(x)
}
\description{Calculate the fold change when x is of size two (always show it >1)}
\value{the proportion of x[1]/x[2] or x[2]/x[1] depending which is larger}
\arguments{\item{x}{input vector of size two}
}

