\name{initialize,DDDataSet-method}
\alias{initialize,DDDataSet-method}
\title{Construct new DDDataSet object...}
\usage{\S4method{initialize}{DDDataSet}(.Object, ..., data=data.frame(), name=paste("Empty name created at",
    date()))
}
\description{Construct new DDDataSet object}
\details{Try to initialise with anything that can be converted to matrix and vectors.}
\arguments{\item{.Object}{the DDDataSet object}
\item{data}{the data slot}
\item{name}{the name slot}
\item{...}{unused}
}

