\name{mcX2CLoop}
\alias{mcX2CLoop}
\title{the inner loop for myX2c is implemented in C...}
\usage{mcX2CLoop(B, numTable, rowSums, colSums)
}
\description{the inner loop for myX2c is implemented in C}
\value{The values of chi-square statistics from random runs}
\arguments{\item{B}{the number of Monte Carlo replicates}
\item{numTable}{the number of conditional tables}
\item{rowSums}{the matrix or row sums for each conditional table (numTables x 4)}
\item{colSums}{the matrix or column sums for each conditional table (numTables x 4)}
}

