\name{operators-CITestResult}
\alias{$,CITestResult-method}
\title{Access slots using the dollar notation...}
\usage{\S4method{$}{CITestResult}(x, name)
\S4method{[[}{CITestResult,ANY,ANY}(x, i, j, ...)
}
\description{Access slots using the dollar notation}
\alias{$,CITestResult-method}
\docType{methods}
\arguments{\item{x}{the CITestResult object}
\item{name}{the slot name}
\item{i}{the slot name}
\item{j}{unused}
\item{...}{unused}
}
\alias{[[,CITestResult,ANY,ANY-method}
\alias{[[,CITestResult-method}

