\name{operators-DDDataSet}
\alias{$,DDDataSet-method}
\title{access a specific variable in the dataset by name...}
\usage{\S4method{$}{DDDataSet}(x, name)
\S4method{[[}{DDDataSet}(x, i, j)
\S4method{[}{DDDataSet,ANY,ANY}(x, i, j, ..., drop=TRUE)
}
\description{access a specific variable in the dataset by name}
\alias{$,DDDataSet-method}
\docType{methods}
\arguments{\item{x}{the DDDataSet object}
\item{name}{the variable name}
\item{i}{variable name}
\item{j}{unusued}
\item{drop}{unused}
\item{...}{unused}
}
\alias{[[,DDDataSet-method}
\alias{[[,DDDataSet-method}
\alias{[,DDDataSet,ANY,ANY-method}
\alias{[,DDDataSet,ANY,ANY,ANY-method}
\alias{[}
\alias{[,DDDataSet-method}

