\name{pcalgMB}
\alias{pcalgMB}
\title{Find the markov blanket for the PC algorithm output...}
\usage{pcalgMB(pc, node)
}
\description{Find the markov blanket for the PC algorithm output}
\value{the inidicies of nodes that constitute the Markov Blanket}
\arguments{\item{pc}{output of PC algorithm from package pcAlgo, object of class "pcAlgo"}
\item{node}{the index of the node for which we are seeking the Markov Blanket}
}

