\name{pcalgNBR}
\alias{pcalgNBR}
\title{Find the neighbourhood for the PC algorithm output...}
\usage{pcalgNBR(pc, node)
}
\description{Find the neighbourhood for the PC algorithm output}
\value{the inidicies of nodes that constitute the (undirected) neighbourhood}
\arguments{\item{pc}{output of PC algorithm from package pcAlgo, object of class "pcAlgo"}
\item{node}{the index of the node for which we are seeking the neighbourhood}
}

