\name{plotSVMPerformance}
\alias{plotSVMPerformance}
\title{Plot SVM performance into a pdf file}
\usage{plotSVMPerformance(obj, results, plot.file)
}
\description{A companion function for \code{svmFeatureSelectionLOOCV()} to plot the results.}
\arguments{\item{obj}{the DDDataSet object for which the SVM performance is measured}
\item{results}{the results from svmFeatureSelectionLOOCV}
\item{plot.file}{the name of the output pdf file}
}

