\name{prob.distr.unif}
\alias{prob.distr.unif}
\title{Uniform distribution for \code{random.bn.fit}}
\usage{prob.distr.unif(n)
}
\description{Uniform distribution function for \code{random.bn.fit}}
\details{Generate 2^n uniformly distributed numbers in range 0 to 1}
\value{vector of 2^n uniform random numbers}
\arguments{\item{n}{number of variables}
}
\examples{# return 8 random uniform numbers
prob.distr.unif(3)}

