\name{rawData,DDDataSet-method}
\alias{rawData,DDDataSet-method}
\alias{rawData}
\title{Raw data.frame with data}
\usage{\S4method{rawData}{DDDataSet}(obj, ...)
}
\description{Return the raw data frame with the variables, and the last column being "class"}
\value{the raw dataframe that contains all the data}
\arguments{\item{obj}{the DDDataSet object}
\item{...}{unused}
}

