\name{variableNames,DDDataSet-method}
\alias{variableNames,DDDataSet-method}
\alias{variableNames}
\title{Names of variables (-class)}
\usage{\S4method{variableNames}{DDDataSet}(obj, ...)
}
\description{Names of variables (without "class")}
\value{only the names of the variables (i.e. without "class")}
\arguments{\item{obj}{the DDDataSet object}
\item{...}{unused}
}

